Move World struct and impl
Re-export world/world.rs in world/mod.rs so paths don't change
